move utils functions to extractimageurls
