+++ development/pkg/extractimageurls/utils_test.go
@@ -1,4 +1,4 @@
-package pkg
+package extractimageurls
 
 import (
 	"reflect"
